Moved the test to the Python parser test file
